+++ tlru.go
@@ -23,7 +23,8 @@ type TLRU[K comparable, V any] struct {
 // tlruConfig represents the configuration of [TLRU]. It should be used with [TLRUOption].
 type tlruConfig struct {
 	lruConfig
-	clock *lruclock.Clock
+	clock   *lruclock.Clock
+	sliding bool
 }
 
 // TLRUOption is used to configure [TLRU] when creating an instance using [NewTTL] constructor.
@@ -44,7 +45,7 @@ func (f LRUOption) apply(c *tlruConfig) error {
 	return f(&c.lruConfig)
 }
 
-// NewTTL creates a [TLRU] instance with the given capacity, expiresAt and options. It creates
+// NewTTL creates a [TLRU] instance with the given capacity, ttl and options. It creates
 // the required [lrucore.TTLCore] instances, initiates the [mux.Mux] for shard routing.
 // It defaults to the Mux with hash/maphash algorithm. Check `tlru/mux` package for alternatives.
 //
@@ -52,7 +53,7 @@ func (f LRUOption) apply(c *tlruConfig) error {
 //
 // Returns [ErrInvalidCapacity] if capacity is not in the range of int32
 // and greater than or equal to twice the number of shards.
-func NewTTL[K comparable, V any](capacity int, expiresAt time.Duration, opts ...TLRUOption) (*TLRU[K, V], error) {
+func NewTTL[K comparable, V any](capacity int, ttl time.Duration, opts ...TLRUOption) (*TLRU[K, V], error) {
 	// build the config
 	cfg := tlruConfig{
 		lruConfig: lruConfig{
@@ -88,7 +89,10 @@ func NewTTL[K comparable, V any](capacity int, expiresAt time.Duration, opts ...
 	}
 
 	createShard := func(cap int) (*lrucore.TTLCore[K, V], error) {
-		return lrucore.NewTTL[K, V](cap, expiresAt, lrucore.WithClock(cfg.clock))
+		if cfg.sliding {
+			return lrucore.NewTTL[K, V](cap, ttl, lrucore.WithClock(cfg.clock), lrucore.WithSliding())
+		}
+		return lrucore.NewTTL[K, V](cap, ttl, lrucore.WithClock(cfg.clock))
 	}
 	cluster, err := assemble(capacity, cfg.shards, hash, createShard)
 	if err != nil {
@@ -113,9 +117,70 @@ func WithClock(clock *lruclock.Clock) tlruOpt {
 	}
 }
 
+// WithSliding enables Sliding TTL on the LRU cache.
+//
+// It will update the timestamp of the key on [TTLCore.Get] and
+// [TTLCore.Put].
+func WithSliding() tlruOpt {
+	return func(c *tlruConfig) error {
+		c.sliding = true
+		return nil
+	}
+}
+
 // Close safely closes the background clock when TTL is enabled on the cache.
 func (l *TLRU[K, V]) Close() {
 	if l.clock != nil {
 		l.clock.Stop()
 	}
+}
+
+// Check [TLRU.Get] on how Get works.
+// It also returns the remaining TTL in the key if it was found in the cache.
+func (l *TLRU[K, V]) GetWithTTL(key K) (V, time.Duration, bool) {
+	shard := l.mux(key)
+	return l.cluster.shards[shard].GetWithTTL(key)
+}
+
+// Check [TLRU.Peek] on how Peek works.
+// It also returns the remaining TTL in the key if it was found in the cache.
+func (l *TLRU[K, V]) PeekWithTTL(key K) (V, time.Duration, bool) {
+	shard := l.mux(key)
+	return l.cluster.shards[shard].PeekWithTTL(key)
+}
+
+// PutWithTTL adds a new value to the cache with the given key and the provided ttl value.
+//
+// Check [TLRU.Put] on how Put works.
+func (l *TLRU[K, V]) PutWithTTL(key K, value V, ttl time.Duration) {
+	shard := l.mux(key)
+	l.cluster.shards[shard].PutWithTTL(key, value, ttl)
+}
+
+// Refresh resets the TTL of an existing key using the default ttl.
+// It returns false if the key could not be found.
+func (l *TLRU[K, V]) Refresh(key K) bool {
+	shard := l.mux(key)
+	return l.cluster.shards[shard].Refresh(key)
+}
+
+// SetTTL resets the TTL of an existing key using the provided ttl value.
+// It returns false if the key could not be found.
+func (l *TLRU[K, V]) SetTTL(key K, ttl time.Duration) bool {
+	shard := l.mux(key)
+	return l.cluster.shards[shard].SetTTL(key, ttl)
+}
+
+// TTL returns the remaining TTL for the key.
+func (l *TLRU[K, V]) TTL(key K) (time.Duration, bool) {
+	shard := l.mux(key)
+	return l.cluster.shards[shard].TTL(key)
+}
+
+// UpsertWithTTL adds a new value to the cache with the given key and the provided ttl value.
+//
+// Check [TLRU.Upsert] on how Upsert works.
+func (l *TLRU[K, V]) UpsertWithTTL(key K, value V, ttl time.Duration) (lrucore.UpsertState, V) {
+	shard := l.mux(key)
+	return l.cluster.shards[shard].UpsertWithTTL(key, value, ttl)
 }